fix and reorganize tests for authentication
